extra fix: failed to subscribe chat with big integer as TG id

+++ hltv_upcoming_events_bot/db/chat.py
@@ -1,18 +1,18 @@
 import logging
-from typing import Optional, Dict, List
+from typing import Optional, List
 
-from sqlalchemy import Boolean, Column, Integer, String, BigInteger
+from sqlalchemy import Column, Integer, String, BigInteger
 from sqlalchemy.exc import NoResultFound
 from sqlalchemy.orm import Session
 
-from hltv_upcoming_events_bot.db.common import Base, get_engine
 from hltv_upcoming_events_bot import domain
+from hltv_upcoming_events_bot.db.common import Base
 
 
 class Chat(Base):
     __tablename__ = "chat"
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, nullable=False)
     title = Column(String)
     type = Column(String)
